Update test data, move helpers
