+++ src/server/services/wmo/status-line-items/wmo-hulpmiddelen.ts
@@ -1,16 +1,14 @@
 import { ZorgnedStatusLineItemTransformerConfig } from '../../zorgned/zorgned-config-and-types';
 
 import {
-  isBeforeToday,
-  isServiceDeliveryActive,
-  isServiceDeliveryStarted,
-} from '../../zorgned/zorgned-helpers';
-import {
   AANVRAAG,
   EINDE_RECHT,
   getTransformerConfigBesluit,
   IN_BEHANDELING,
-  isServiceDeliveryDecisionActive,
+  isBeforeToday,
+  isDecisionWithDeliveryActive,
+  isServiceDeliveryActive,
+  isServiceDeliveryStarted,
   MEER_INFORMATIE,
 } from './wmo-generic';
 
@@ -18,7 +16,7 @@ export const hulpmiddelen: ZorgnedStatusLineItemTransformerConfig[] = [
   AANVRAAG,
   IN_BEHANDELING,
   MEER_INFORMATIE,
-  getTransformerConfigBesluit(isServiceDeliveryDecisionActive, true),
+  getTransformerConfigBesluit(isDecisionWithDeliveryActive, true),
   {
     status: 'Opdracht gegeven',
     datePublished: (aanvraag) => aanvraag.datumOpdrachtLevering ?? '',
