+++ src/server/services/zorgned/zorgned-helpers.ts
@@ -1,5 +1,3 @@
-import { isSameDay, parseISO } from 'date-fns';
-import { isDateInPast } from '../../../universal/helpers/date';
 import {
   TextPartContents,
   ZorgnedAanvraagTransformed,
@@ -18,39 +16,4 @@ export function parseLabelContent(
   }
 
   return rText;
-}
-
-export function isBeforeToday(dateStr: string | null, compareDate: Date) {
-  if (!dateStr) {
-    return false;
-  }
-  return isSameDay(parseISO(dateStr), compareDate)
-    ? false
-    : isDateInPast(dateStr, compareDate);
-}
-
-export function isServiceDeliveryStarted(
-  sourceData: ZorgnedAanvraagTransformed,
-  compareDate: Date
-) {
-  return isBeforeToday(sourceData.datumBeginLevering, compareDate);
-}
-
-export function isServiceDeliveryStopped(
-  sourceData: ZorgnedAanvraagTransformed,
-  compareDate: Date
-) {
-  return isBeforeToday(sourceData.datumEindeLevering, compareDate);
-}
-
-export function isServiceDeliveryActive(
-  sourceData: ZorgnedAanvraagTransformed,
-  compareDate: Date
-) {
-  return (
-    sourceData.isActueel &&
-    isServiceDeliveryStarted(sourceData, compareDate) &&
-    !isServiceDeliveryStopped(sourceData, compareDate) &&
-    !isBeforeToday(sourceData.datumEindeGeldigheid, compareDate)
-  );
 }
