+++ src/universal/config/feature-toggles.ts
@@ -46,5 +46,6 @@ export const FeatureToggle = {
   wiorDatasetActive: true,
   wiorMeldingen: true,
   zorgnedDocumentAttachmentsActive: true,
+  zorgnedDocumentDecisionDateActive: !IS_PRODUCTION,
   zorgv2ThemapaginaActive: !IS_PRODUCTION,
 };
